Production check: fix 20 confirmed MEDIUM+ findings (iter 1+2 adversarially verified)
C++ fixes (8 findings):
- my_kriging_weights.h: Fix signed int overflow in matrix indexing (F-C01)
- api_helpers.cpp: Pre-multiplication overflow check in get_shape_volume (F-H01)
- indicator_kriging.h: Add OpenMP cancellation support (F-M14)
- sugarbox_grid.cpp: Add int32 overflow guard for large grids (F-M15)
- simple_cokriging_markI.cpp: Change build_system void->bool return (F-M18)
- median_ik.cpp: Add OpenMP cancellation support (I2-07)
- variograms.cpp/ellipsoid.cpp/stack_layers.cpp: Add try/catch to 9 CVAR_API extern C functions (I2-01)
- variograms.cpp: Replace non-thread-safe rand() with std::mt19937 (I2-03)

Python fixes (12 findings):
- geo.py: Add isinstance guards preventing buffer overflow via @accepts_tuple (F-M35, I2-10, I2-11, I2-12)
- geo.py: Add 3D shape validation to _create_hpgl_ind_masked_array (I2-14)
- geo.py: Add fix_shape method to IndProperty (I2-16)
- geo.py: Add type guards in median_ik/indicator_kriging (I2-18)
- validation.py: Change validate_seed from warning to error (F-H08)
- routines.py: Add empty dict guards in SaveGSLIBPointSet/Cubes2PointSet (F-M41, I2-15)
- sgs.py: Add isinstance guard preventing buffer overflow (I2-09)
- sis.py: Add isinstance guard for indicator_count (I2-18)
- variogram.py: Fix quadratic O(L^2) vstack to O(L) accumulate (I2-17)

Build fixes:
- select.h: Add template overloads for mean providers without .size()
- property_array.h: Add size_type/size() to indicator_array_adapter_t

Test results: 829 passed, 2 skipped, 1 xfailed
Lint: clean
Mypy: 41 pre-existing union-attr warnings (unchanged)

+++ src/geo_bsd/hpgl/my_kriging_weights.h
@@ -143,8 +143,8 @@ namespace hpgl
 		{
 			for (int j = i, end_j = end_i; j < end_j; ++j)
 			{
-				A[i*size + j] = covariances(coords[i], coords[j]);
-				A[j*size + i] = A[i*size + j];
+				A[static_cast<size_t>(i) * size + j] = covariances(coords[i], coords[j]);
+				A[static_cast<size_t>(j) * size + i] = A[static_cast<size_t>(i) * size + j];
 			}
 			b[i] = covariances(coords[i], center_coord);
 			// b2 preserves original RHS values for variance calculation;
@@ -318,8 +318,8 @@ namespace hpgl
 		{
 			for (int j = i, end_j = end_i; j < end_j; ++j)
 			{
-				ws.A[i*size + j] = covariances(coords[i], coords[j]);
-				ws.A[j*size + i] = ws.A[i*size + j];
+				ws.A[static_cast<size_t>(i) * size + j] = covariances(coords[i], coords[j]);
+				ws.A[static_cast<size_t>(j) * size + i] = ws.A[static_cast<size_t>(i) * size + j];
 			}
 			ws.b[i] = covariances(coords[i], center_coord);
 			ws.b2[i] = ws.b[i];
@@ -463,8 +463,8 @@ namespace hpgl
 		{
 			for (int j = i, end_j = end_i; j < end_j; ++j)
 			{
-				A[i*size + j] = covariances(coords[i], coords[j]);
-				A[j*size + i] = A[i*size + j];
+				A[static_cast<size_t>(i) * size + j] = covariances(coords[i], coords[j]);
+				A[static_cast<size_t>(j) * size + i] = A[static_cast<size_t>(i) * size + j];
 			}
 			b[i] = covariances(coords[i], center);
 			// b2 preserves original RHS values for variance calculation
@@ -755,8 +755,8 @@ namespace hpgl
 		{
 			for (int j = i, end_j = end_i; j < end_j; ++j)
 			{
-				ws.A[i*size + j] = covariances(coords[i], coords[j]);
-				ws.A[j*size + i] = ws.A[i*size + j];
+				ws.A[static_cast<size_t>(i) * size + j] = covariances(coords[i], coords[j]);
+				ws.A[static_cast<size_t>(j) * size + i] = ws.A[static_cast<size_t>(i) * size + j];
 			}
 			ws.b[i] = covariances(coords[i], center);
 			ws.b2[i] = ws.b[i];
@@ -1066,9 +1066,9 @@ namespace hpgl
 		{
 			for (int j = i; j < size; ++j)
 			{
-				A[i * size + j] =
+				A[static_cast<size_t>(i) * size + j] =
 					cov(coords[i], coords[j]) * (sigmas[i] * sigmas[j]);
-				A[j * size + i] = A[i * size + j];
+				A[static_cast<size_t>(j) * size + i] = A[static_cast<size_t>(i) * size + j];
 			}
 			b[i] = cov(coords[i], center) * (sigmas[i] * sigmac);
 		}
@@ -1232,9 +1232,9 @@ namespace hpgl
 		{
 			for (int j = i; j < size; ++j)
 			{
-				ws.A[i * size + j] =
+				ws.A[static_cast<size_t>(i) * size + j] =
 					cov(coords[i], coords[j]) * (ws.sigmas[i] * ws.sigmas[j]);
-				ws.A[j * size + i] = ws.A[i * size + j];
+				ws.A[static_cast<size_t>(j) * size + i] = ws.A[static_cast<size_t>(i) * size + j];
 			}
 			ws.b[i] = cov(coords[i], center) * (ws.sigmas[i] * sigmac);
 		}
